+++ biblioteca-4/tablas/db-obras/buscar-1.php
@@ -27,15 +27,18 @@ if (muestraFormulario()) {
     print "        <tbody>\n";
     print "          <tr>\n";
     print "            <td>Autor:</td>\n";
-    print "            <td><input type=\"text\" name=\"autor\" size=\"$db[tamObrasAutor]\" maxlength=\"$db[tamObrasAutor]\" autofocus></td>\n";
+    print "            <td><input type=\"text\" name=\"autor\" size=\"$db[tamObrasAutor]\" maxlength=\"$db[tamObrasAutor]\""
+        . imprimeAvisosIndividuales("obras", "buscar-2", "autor", "valor") . " autofocus></td>\n";
     print "          </tr>\n";
     print "          <tr>\n";
     print "            <td>Título:</td>\n";
-    print "            <td><input type=\"text\" name=\"titulo\" size=\"$db[tamObrasTitulo]\" maxlength=\"$db[tamObrasTitulo]\"></td>\n";
+    print "            <td><input type=\"text\" name=\"titulo\" size=\"$db[tamObrasTitulo]\" maxlength=\"$db[tamObrasTitulo]\""
+        . imprimeAvisosIndividuales("obras", "buscar-2", "titulo", "valor") . "></td>\n";
     print "          </tr>\n";
     print "          <tr>\n";
     print "            <td>Editorial:</td>\n";
-    print "            <td><input type=\"text\" name=\"editorial\" size=\"$db[tamObrasEditorial]\" maxlength=\"$db[tamObrasEditorial]\"></td>\n";
+    print "            <td><input type=\"text\" name=\"editorial\" size=\"$db[tamObrasEditorial]\" maxlength=\"$db[tamObrasEditorial]\""
+        . imprimeAvisosIndividuales("obras", "buscar-2", "editorial", "valor") . "></td>\n";
     print "          </tr>\n";
     print "        </tbody>\n";
     print "      </table>\n";